# specs/001-build-cloudyintel-website/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CloudyIntel Run Orchestration API
  version: 0.1.0
  description: API contract for query submission, provider run tracking, validation evidence, and workflow state streaming.
servers:
  - url: http://localhost:8000
tags:
  - name: Runs
  - name: History
  - name: Workflow
paths:
  /api/v1/runs:
    post:
      tags: [Runs]
      summary: Create a new architecture query run
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '202':
          description: Run accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRunResponse'
  /api/v1/runs/{runId}:
    get:
      tags: [Runs]
      summary: Get aggregated run status and latest outputs
      operationId: getRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetail'
        '404':
          description: Not found
  /api/v1/runs/{runId}/providers/{provider}/findings:
    get:
      tags: [Runs]
      summary: List validation findings for a provider run
      operationId: listFindings
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/Provider'
      responses:
        '200':
          description: Findings list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidationFinding'
  /api/v1/runs/{runId}/providers/{provider}/events/stream:
    get:
      tags: [Workflow]
      summary: Stream provider workflow events (SSE)
      operationId: streamWorkflowEvents
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/Provider'
      responses:
        '200':
          description: text/event-stream
          content:
            text/event-stream:
              schema:
                type: string
  /api/v1/history:
    get:
      tags: [History]
      summary: List persisted run history for current workspace
      operationId: listHistory
      parameters:
        - in: query
          name: provider
          required: false
          schema:
            $ref: '#/components/schemas/Provider'
      responses:
        '200':
          description: History records
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/HistoryItem'
components:
  parameters:
    RunId:
      in: path
      name: runId
      required: true
      schema:
        type: string
        format: uuid
    Provider:
      in: path
      name: provider
      required: true
      schema:
        $ref: '#/components/schemas/Provider'
  schemas:
    Provider:
      type: string
      enum: [aws, azure]
    ProviderScope:
      type: string
      enum: [aws, azure, both]
    CreateRunRequest:
      type: object
      required: [query, provider_scope]
      properties:
        query:
          type: string
          minLength: 1
        provider_scope:
          $ref: '#/components/schemas/ProviderScope'
        constraints:
          type: object
          additionalProperties: true
        optimization_preferences:
          type: object
          additionalProperties: true
    CreateRunResponse:
      type: object
      required: [run_id, status]
      properties:
        run_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running]
    RunDetail:
      type: object
      required: [run_id, provider_runs]
      properties:
        run_id:
          type: string
          format: uuid
        query:
          type: string
        provider_scope:
          $ref: '#/components/schemas/ProviderScope'
        provider_runs:
          type: array
          items:
            $ref: '#/components/schemas/ProviderRunDetail'
    ProviderRunDetail:
      type: object
      required: [provider, status, iteration_count, accepted]
      properties:
        provider:
          $ref: '#/components/schemas/Provider'
        status:
          type: string
          enum: [generating, validating, optimizing, accepted, stopped, failed]
        iteration_count:
          type: integer
          minimum: 0
          maximum: 3
        accepted:
          type: boolean
        terminal_reason:
          type: string
          enum: [accepted, iteration_limit, evidence_unavailable, domain_failure, timeout]
        unresolved_findings:
          type: integer
          minimum: 0
    ValidationFinding:
      type: object
      required: [finding_id, domain, outcome, reason]
      properties:
        finding_id:
          type: string
          format: uuid
        iteration:
          type: integer
          minimum: 0
          maximum: 3
        domain:
          type: string
          enum: [compute, storage, network, database, cross_domain]
        outcome:
          type: string
          enum: [pass, fail, warn]
        severity:
          type: string
          enum: [critical, high, medium, low]
        reason:
          type: string
        remediation:
          type: string
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceReference'
    EvidenceReference:
      type: object
      required: [source_uri, title, availability_status]
      properties:
        source_uri:
          type: string
          format: uri
        title:
          type: string
        excerpt:
          type: string
        availability_status:
          type: string
          enum: [available, unavailable]
    HistoryItem:
      type: object
      required: [run_id, query, provider_scope, created_at]
      properties:
        run_id:
          type: string
          format: uuid
        query:
          type: string
        provider_scope:
          $ref: '#/components/schemas/ProviderScope'
        created_at:
          type: string
          format: date-time
        provider_runs:
          type: array
          items:
            $ref: '#/components/schemas/ProviderRunDetail'